removed concurrency from RunCommandThatMustSucceed
